Fibonacci iOS app. Shows unlimited number of rows.
Andrey Kasatkin
December 10th, 2014
TodayTix coding challenge

THE CHALLENGE:
Write an iOS app to display the fibonacci sequence in a UITableView.
Please be aware of the amount of NSOperations done and give the most
efficient solution.

ABOUT:
This app displays the Fibonacci numbers in a UITableView from Fib(1) =
1 to Fib(500 - this number can be changed). The caluclation of Fibs
after 93 are processed on the background due to the different method of
computetion when out of bounds for unsigned long long int.